refactor(ipc): 将 IPC 通信从回调模式重构为 async/await 模式
- 将主进程和渲染进程之间的 IPC 通信从基于事件的回调模式重构为基于 `Promise` 的 `async/await` 模式。主要变更包括：
- 将 `ipc.send`/`ipc.on` 模式改为 `ipc.invoke`/`ipc.handle` 模式
- 重构文件操作相关的 IPC 通信（打开文件、选择路径等）
- 重构模板操作相关的 IPC 通信（加载、删除、重命名等）
- 拆分 `IOManager` 为 `settingManager` 和 `templateManager` 两个独立模块
- 将窗口管理相关的 IPC 处理移至 `windowManager` 模块
- 优化代码结构，提高可维护性和可读性
